requirements.txt
pymongo==4.10.1
python-dotenv==0.20.0
python_telegram_bot==20.0a2
telegram==0.0.1
dnspython==2.2.1